creating separate folder for allocation managers
