+++ src/main/scala/beam/agentsim/agents/rideHail/allocationManagers/StanfordRideHailAllocationManagerV1.scala
@@ -1,11 +1,10 @@
-package beam.agentsim.agents.rideHail
+package beam.agentsim.agents.rideHail.allocationManagers
 
-import beam.agentsim.agents.rideHail.RideHailingManager.{RideHailingAgentLocation, RideHailingInquiry}
-import beam.agentsim.events.SpaceTime
+import beam.agentsim.agents.rideHail.RideHailingManager
 import beam.router.BeamRouter.Location
-import beam.router.RoutingModel
-import org.matsim.api.core.v01.{Coord, Id}
+import org.matsim.api.core.v01.Id
 import org.matsim.vehicles.Vehicle
+
 import scala.util.control.Breaks._
 
 /*
